Data.Type.Quantification

Existential Quantification

data Any :: (k ~> Type) -> f k -> Type where

An Any p as is a witness that, for at least one item a in the type-level collection as, the predicate p a is true.

Constructors

Any :: Elem f as a -> (p @@ a) -> Any p as

entailAny

Arguments

:: (forall a. Elem f as a -> (p @@ a) -> q @@ a)
implication

-> Any p as

-> Any q as

If there exists an a s.t. p a, and if p implies q, then there must exist an a s.t. q a.

entailAnyF

Arguments

:: Functor h

=> (forall a. Elem f as a -> (p @@ a) -> h (q @@ a))
implication in context

-> Any p as

-> h (Any q as)

If p implies q under some context h, and if there exists some a such that p a, then there must exist some a such that p q under that context h.

h might be something like, say, Maybe, to give predicate that is either provably true or unprovably false.

Note that it is not possible to do this with p a -> Decision (q a). This is if the p a -> Decision (q a) implication is false, there it doesn't mean that there is no a such that q a, necessarily. There could have been an a where p does not hold, but q does.

entailAny'

Arguments

:: forall (as :: f k). (forall a. (p @@ a) -> q @@ a)
implication

-> Any p as

-> Any q as

entailAny, but without the membership witness.

entailAnyF'

Arguments

:: forall (as :: f k). Functor h

=> (forall a. (p @@ a) -> h (q @@ a))
implication in context

-> Any p as

-> h (Any q as)

entailAnyF, but without the membership witness.

Universal Quantification

newtype All p (as :: f k)

An All p as is a witness that, the predicate p a is true for all items a in the type-level collection as.

Constructors

All

Fields

runAll :: forall a. Elem f as a -> p @@ a

entailAll

Arguments

:: (forall a. Elem f as a -> (p @@ a) -> q @@ a)
implication

-> All p as

-> All q as

If for all a we have p a, and if p implies q, then for all a we must also have p a.

entailAllF

Arguments

:: (Universe f, Applicative h, SingI as)

=> (forall a. Elem f as a -> (p @@ a) -> h (q @@ a))
implication in context

-> All p as

-> h (All q as)

If p implies q under some context h, and if we have p a for all a, then we must have q a for all a under context h.

decideEntailAll

Arguments

:: (Universe f, SingI as)

=> (forall a. Elem f as a -> (p @@ a) -> Decision (q @@ a))
decidable implication

-> All p as

-> Decision (All q as)

If we have p a for all a, and p a can be used to test for q a, then we can test all as for q a.

entailAll'

Arguments

:: forall (as :: f k). (forall a. (p @@ a) -> q @@ a)
implication

-> All p as

-> All q as

entailAll, but without the membership witness.

entailAllF'

Arguments

:: forall (as :: f k). (Universe f, Applicative h, SingI as)

=> (forall a. (p @@ a) -> h (q @@ a))
implication in context

-> All p as

-> h (All q as)

entailAllF, but without the membership witness.

decideEntailAll'

Arguments

:: forall (as :: f k). (Universe f, SingI as)

=> (forall a. (p @@ a) -> Decision (q @@ a))
decidable implication

-> All p as

-> Decision (All q as)

decideEntailAll, but without the membeship witness.