Data.Type.Universe

Documentation

data family Elem (f :: Type -> Type) :: f k -> k -> Type

A witness for membership of a given item in a type-level collection

class Universe (f :: Type -> Type) where

Typeclass for a type-level container that you can quantify or lift type-level predicates over.

Minimal complete definition

idecideAny, idecideAll, igenAllA

Methods

idecideAny

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (Any f p as)
predicate on collection

You should read this type as:

decideAny' :: (Sing a  -> Decision (p a)    )
           -> (Sing as -> Decision (Any p as)

It lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if any item in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests existential quantification.

idecideAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (All f p as)
predicate on collection

You should read this type as:

decideAll' :: (Sing a  -> Decision (p a)    )
           -> (Sing as -> Decision (All p as)

It lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if all items in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests universal quantification.

igenAllA

Arguments

:: forall (p :: k ~> Type) (as :: f k). Applicative h

=> (forall a. Elem f as a -> Sing a -> h (p @@ a))
predicate on value in context

-> Sing as -> h (All f p as)
predicate on collection in context

If p a is true for all values a in as under some (Applicative) context h, then you can create an All p as under that Applicative context h.

Can be useful with Identity (which is basically unwrapping and wrapping All), or with Maybe (which can express predicates that are either provably true or not provably false).

decideAny

Arguments

:: forall (p :: k ~> Type). Universe f

=> Test p
predicate on value

-> Test (TyCon1 (Any f p))
predicate on collection

decideAll

Arguments

:: forall (p :: k ~> Type). Universe f

=> Test p
predicate on value

-> Test (TyCon1 (All f p))
predicate on collection

genAllA

Arguments

:: forall (p :: k ~> Type) (as :: f k). (Universe f, Applicative h)

=> (forall a. Sing a -> h (p @@ a))
predicate on value in context

-> Sing as -> h (All f p as)
predicate on collection in context

igenAllA, but without the membership witness.

genAll

Arguments

:: forall (p :: k ~> Type). Universe f

=> Given p
always-true predicate on value

-> Given (TyCon1 (All f p))
always-true predicate on collection

igenAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). Universe f

=> (forall a. Elem f as a -> Sing a -> p @@ a)
always-true predicate on value

-> Sing as -> All f p as
always-true predicate on collection

If p a is true for all values a in as, then we have All p as. Basically witnesses the definition of All.

foldMapUni :: forall f k (as :: f k) m. (Universe f, SingI as, Monoid m) => (forall (a :: k). Sing a -> m) -> Sing as -> m

A foldMap over all items in a collection.

ifoldMapUni :: forall f k (as :: f k) m. (Universe f, SingI as, Monoid m) => (forall a. Elem f as a -> Sing a -> m) -> Sing as -> m

foldMapUni but with access to the index.

select

Arguments

:: Universe f

=> Elem f as a
Witness

-> Sing as
Collection

-> Sing a

Extract the item from the container witnessed by the Elem

pickElem :: forall f k (as :: f k) a. (Universe f, SingI as, SingI a, SDecide k) => Decision (Elem f as a)

Automatically generate a witness for a member, if possible

newtype All f p (as :: f k)

An All p as is a witness that, the predicate p a is true for all items a in the type-level collection as.

Constructors

All

Fields

runAll :: forall a. Elem f as a -> p @@ a

data Any f :: (k ~> Type) -> f k -> Type where

An Any p as is a witness that, for at least one item a in the type-level collection as, the predicate p a is true.

Constructors

Any :: Elem f as a -> (p @@ a) -> Any f p as