Data.Type.Elem

Documentation

type family Elem (f :: Type -> Type) :: f k -> k -> Type

A witness for membership of a given item in a type-level collection

class Universe (f :: Type -> Type) where

Typeclass for a type-level container that you can quantify or lift type-level predicates over.

Minimal complete definition

decideAny, decideAll, genAllA

Methods

decideAny

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (Any p as)
predicate on collection

You should read this type as:

decideAny :: (Sing a  -> Decision (p a)    )
          -> (Sing as -> Decision (Any p as)

It lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if any item in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests existential quantification.

decideAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (All p as)
predicate on collection

You should read this type as:

decideAll :: (Sing a  -> Decision (p a)    )
          -> (Sing as -> Decision (All p as)

It lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if all items in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests universal quantification.

genAllA

Arguments

:: forall (p :: k ~> Type) (as :: f k). Applicative h

=> (forall a. Elem f as a -> Sing a -> h (p @@ a))
predicate on value in context

-> Sing as -> h (All p as)
predicate on collection in context

If p a is true for all values a in as under some (Applicative) context h, then you can create an All p as under that Applicative context h.

Can be useful with Identity (which is basically unwrapping and wrapping All), or with Maybe (which can express predicates that are either provably true or not provably false).

genAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). Universe f

=> (forall a. Elem f as a -> Sing a -> p @@ a)
always-true predicate on value

-> Sing as -> All p as
always-true predicate on collection

If p a is true for all values a in as, then we have All p as. Basically witnesses the definition of All.

select

Arguments

:: Universe f

=> Elem f as a
Witness

-> Sing as
Collection

-> Sing a

Extract the item from the container witnessed by the Elem

data Any :: (k ~> Type) -> f k -> Type where

An Any p as is a witness that, for at least one item a in the type-level collection as, the predicate p a is true.

Constructors

Any :: Elem f as a -> (p @@ a) -> Any p as

newtype All p (as :: f k)

An All p as is a witness that, the predicate p a is true for all items a in the type-level collection as.

Constructors

All

Fields

runAll :: forall a. Elem f as a -> p @@ a

entailAny

Arguments

:: forall (as :: f k). (forall a. Elem f as a -> (p @@ a) -> q @@ a)
implication

-> Any p as

-> Any q as

If there exists an a s.t. p a, and if p implies q, then there must exist an a s.t. q a.

entailAnyF

Arguments

:: Functor h

=> (forall a. Elem f as a -> (p @@ a) -> h (q @@ a))
implication in context

-> Any p as

-> h (Any q as)

If p implies q under some context h, and if there exists some a such that p a, then there must exist some a such that p q under that context h.

h might be something like, say, Maybe, to give predicate that is either provably true or unprovably false.

Note that it is not possible to do this with p a -> Decision (q a). This is if the p a -> Decision (q a) implication is false, there it doesn't mean that there is no a such that q a, necessarily. There could have been an a where p does not hold, but q does.

entailAll

Arguments

:: forall (as :: f k). (forall a. Elem f as a -> (p @@ a) -> q @@ a)
implication

-> All p as

-> All q as

If for all a we have p a, and if p implies q, then for all a we must also have p a.

entailAllF

Arguments

:: (Universe f, Applicative h, SingI as)

=> (forall a. Elem f as a -> (p @@ a) -> h (q @@ a))
implication in context

-> All p as

-> h (All q as)

If p implies q under some context h, and if we have p a for all a, then we must have q a for all a under context h.

decideEntailAll

Arguments

:: forall (as :: f k). (Universe f, SingI as)

=> (forall a. Elem f as a -> (p @@ a) -> Decision (q @@ a))
decidable implication

-> All p as

-> Decision (All q as)

If we have p a for all a, and p a can be used to test for q a, then we can test all as for q a.

data Index :: [k] -> k -> Type where

Witness an item in a type-level list by providing its index.

Constructors

IZ :: Index (a ': as) a

IS :: Index bs a -> Index (b ': bs) a

data IsJust :: Maybe k -> k -> Type where

Witness an item in a type-level Maybe by proving the Maybe is Just.

Constructors

IsJust :: IsJust (Just a) a

data NEIndex :: NonEmpty k -> k -> Type where

Witness an item in a type-level NonEmpty by either indicating that it is the "head", or by providing an index in the "tail".

Constructors

NEHead :: NEIndex (a :| as) a

NETail :: Index as a -> NEIndex (b :| as) a

data Snd :: (j, k) -> k -> Type where

Trivially witness an item in the second field of a type-level tuple.

Constructors

Snd :: Snd '(a, b) b