Data.Type.Universe

Documentation

class Universe (f :: Type -> Type) where

Typeclass for a type-level container that you can quantify or lift type-level predicates over.

Minimal complete definition

decideAny, decideAll, genAllA

Methods

decideAny

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (Any p as)
predicate on collection

You should read this type as:

decideAny :: (Sing a  -> Decision (p a)    )
          -> (Sing as -> Decision (Any p as)

It lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if any item in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests existential quantification.

decideAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). (forall a. Elem f as a -> Sing a -> Decision (p @@ a))
predicate on value

-> Sing as -> Decision (All p as)
predicate on collection

You should read this type as:

decideAll :: (Sing a  -> Decision (p a)    )
          -> (Sing as -> Decision (All p as)

It lifts a predicate p on an individual a into a predicate that on a collection as that is true if and only if all items in as satisfies the original predicate.

That is, it turns a predicate of kind k ~> Type into a predicate of kind f k ~> Type.

Essentially tests universal quantification.

genAllA

Arguments

:: forall (p :: k ~> Type) (as :: f k). Applicative h

=> (forall a. Elem f as a -> Sing a -> h (p @@ a))
predicate on value in context

-> Sing as -> h (All p as)
predicate on collection in context

If p a is true for all values a in as under some (Applicative) context h, then you can create an All p as under that Applicative context h.

Can be useful with Identity (which is basically unwrapping and wrapping All), or with Maybe (which can express predicates that are either provably true or not provably false).

genAll

Arguments

:: forall (p :: k ~> Type) (as :: f k). Universe f

=> (forall a. Elem f as a -> Sing a -> p @@ a)
always-true predicate on value

-> Sing as -> All p as
always-true predicate on collection

If p a is true for all values a in as, then we have All p as. Basically witnesses the definition of All.

select

Arguments

:: Universe f

=> Elem f as a
Witness

-> Sing as
Collection

-> Sing a

Extract the item from the container witnessed by the Elem

splitSing :: forall f (as :: f k). Universe f => Sing as -> All (TyCon1 Sing) as

Split a Sing as into a proof that all a in as exist.

data Index :: [k] -> k -> Type where

Witness an item in a type-level list by providing its index.

Constructors

IZ :: Index (a ': as) a

IS :: Index bs a -> Index (b ': bs) a

data IsJust :: Maybe k -> k -> Type where

Witness an item in a type-level Maybe by proving the Maybe is Just.

Constructors

IsJust :: IsJust (Just a) a

data NEIndex :: NonEmpty k -> k -> Type where

Witness an item in a type-level NonEmpty by either indicating that it is the "head", or by providing an index in the "tail".

Constructors

NEHead :: NEIndex (a :| as) a

NETail :: Index as a -> NEIndex (b :| as) a

data Snd :: (j, k) -> k -> Type where

Trivially witness an item in the second field of a type-level tuple.

Constructors

Snd :: Snd '(a, b) b

Orphan instances

Universe []

Instance details

Methods

decideAny :: (forall (a :: k). Elem [] as a -> Sing a -> Decision (p @@ a)) -> Sing as -> Decision (Any p as)

decideAll :: (forall (a :: k). Elem [] as a -> Sing a -> Decision (p @@ a)) -> Sing as -> Decision (All p as)

genAllA :: Applicative h => (forall (a :: k). Elem [] as a -> Sing a -> h (p @@ a)) -> Sing as -> h (All p as)

Universe Maybe

Instance details

Methods

decideAny :: (forall (a :: k). Elem Maybe as a -> Sing a -> Decision (p @@ a)) -> Sing as -> Decision (Any p as)

decideAll :: (forall (a :: k). Elem Maybe as a -> Sing a -> Decision (p @@ a)) -> Sing as -> Decision (All p as)

genAllA :: Applicative h => (forall (a :: k). Elem Maybe as a -> Sing a -> h (p @@ a)) -> Sing as -> h (All p as)

Universe NonEmpty

Instance details

Methods

decideAny :: (forall (a :: k). Elem NonEmpty as a -> Sing a -> Decision (p @@ a)) -> Sing as -> Decision (Any p as)

decideAll :: (forall (a :: k). Elem NonEmpty as a -> Sing a -> Decision (p @@ a)) -> Sing as -> Decision (All p as)

genAllA :: Applicative h => (forall (a :: k). Elem NonEmpty as a -> Sing a -> h (p @@ a)) -> Sing as -> h (All p as)

Universe ((,) j)

Instance details

Methods

decideAny :: (forall (a :: k). Elem ((,) j) as a -> Sing a -> Decision (p @@ a)) -> Sing as -> Decision (Any p as)

decideAll :: (forall (a :: k). Elem ((,) j) as a -> Sing a -> Decision (p @@ a)) -> Sing as -> Decision (All p as)

genAllA :: Applicative h => (forall (a :: k). Elem ((,) j) as a -> Sing a -> h (p @@ a)) -> Sing as -> h (All p as)